+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 - Fixes warning about missing viewer prop - ash
 - Fixes warning about duplicate keys in list - ash
+- Fixes top buttons re-appearing behind city picker - ash
 
 ### 1.8.14